removing warnings
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@5340 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/netload/NLEdgeControlBuilder.h
@@ -133,10 +133,10 @@ public:
 
 protected:
     /// @brief A running numer for lane numbering
-    size_t myCurrentNumericalLaneID;
+    unsigned int myCurrentNumericalLaneID;
 
     /// @brief A running numer for edge numbering
-    size_t myCurrentNumericalEdgeID;
+    unsigned int myCurrentNumericalEdgeID;
 
     /** @brief storage for edges
      *
